+++ src/domain/models/partner/IPartnerRepository.ts
@@ -3,4 +3,5 @@ import { Partner } from './partner';
 export interface IPartnerRepository {
   save(item: Partner): Promise<Partner>;
   findOne(id: string): Promise<Partner>;
+  findNearestBy(coordinates: number[]): Promise<Partner>;
 }