+++ src/infrastructure/mongodb/repositories/PartnerRepository.ts
@@ -1,13 +1,14 @@
 import { DuplicateKeyException } from '../../../domain/exceptions/DuplicateKeyException';
 import { NotFoundPartnerException } from '../../../domain/exceptions/NotFoundPartnerException';
 import { SavePartnerException } from '../../../domain/exceptions/SavePartnerException';
-import { Address } from '../../../domain/models/partner/address';
-import { CoverageArea } from '../../../domain/models/partner/CoverageArea';
 import { GeojsonType } from '../../../domain/models/partner/geojsonType';
 import { IPartnerRepository } from '../../../domain/models/partner/IPartnerRepository';
 import { Partner } from '../../../domain/models/partner/partner';
+import { FindNearest } from '../../geolib/findNearest';
 import { MongoHelper } from '../../utils/MongoHelper';
+import { Util } from '../../utils/util';
 import { CodeException } from '../CodeException';
+import { PartnerDBModelMapper } from '../mappers/PartnerDBModelMapper';
 import { PartnerDBModel } from '../models/PartnerDBModel';
 import { Repository } from './Repository';
 
@@ -26,10 +27,10 @@ export class PartnerRepository implements IPartnerRepository {
 
   public async save(partner: Partner): Promise<Partner> {
     try {
-      const partnerDBModel = this.toDbModel(partner);
+      const partnerDBModel = PartnerDBModelMapper.toDbModel(partner);
       const result = await this.repository.save(partnerDBModel);
 
-      return this.toDomain(result);
+      return PartnerDBModelMapper.toDomain(result);
     } catch (error) {
       if (error.code === CodeException.DuplicateKeyException) {
         throw new DuplicateKeyException(error.message);
@@ -47,7 +48,7 @@ export class PartnerRepository implements IPartnerRepository {
         throw new NotFoundPartnerException('Partner id not found');
       }
 
-      return this.toDomain(partnerDBModel);
+      return PartnerDBModelMapper.toDomain(partnerDBModel);
     } catch (error) {
       if (error instanceof NotFoundPartnerException) {
         throw error;
@@ -57,32 +58,37 @@ export class PartnerRepository implements IPartnerRepository {
     }
   }
 
-  private toDbModel(partner: Partner): PartnerDBModel {
-    return {
-      tradingName: partner.getTradingName(),
-      ownerName: partner.getOwnerName(),
-      document: partner.getDocumentNumber(),
-      coverageArea: {
-        type: partner.getCoverageAreaType(),
-        coordinates: partner.getCoverageCoordinates(),
-      },
-      address: {
-        type: partner.getAddressType(),
-        coordinates: partner.getAddressCoordinates(),
-      },
-    };
+  public async findNearestBy(coordinates: number[]): Promise<Partner> {
+    try {
+      const partnersWithinCoordinates = await this.repository.find(this.createFilter(coordinates));
+
+      if (Util.isEmpty(partnersWithinCoordinates)) {
+        throw new NotFoundPartnerException('No partner found in this area');
+      }
+
+      const point = { latitude: coordinates[Partner.LATITUDE], longitude: coordinates[Partner.LONGITUDE] };
+      const nearestPartner = FindNearest.from(point, partnersWithinCoordinates);
+
+      return PartnerDBModelMapper.toDomain(nearestPartner);
+    } catch (error) {
+      if (error instanceof NotFoundPartnerException) {
+        throw error;
+      }
+
+      throw new SavePartnerException(error.message);
+    }
   }
 
-  private toDomain(partnerDBModel: PartnerDBModel): Partner {
-    return Partner.builder()
-      .withId(partnerDBModel._id)
-      .withTradingName(partnerDBModel.tradingName)
-      .withOwnerName(partnerDBModel.ownerName)
-      .withDocument(partnerDBModel.document)
-      .withCoverageArea(
-        CoverageArea.builder().withType(GeojsonType[partnerDBModel.address.type]).withCoordinates(partnerDBModel.coverageArea.coordinates).build()
-      )
-      .withAddress(Address.builder().withType(GeojsonType[partnerDBModel.coverageArea.type]).withCoordinates(partnerDBModel.address.coordinates).build())
-      .build();
+  private createFilter(coordinates: number[]) {
+    return {
+      coverageArea: {
+        $geoIntersects: {
+          $geometry: {
+            type: GeojsonType.Point,
+            coordinates: coordinates,
+          },
+        },
+      },
+    };
   }
 }