+++ src/infrastructure/utils/util.ts
@@ -1,0 +1,5 @@
+export class Util {
+  public static isEmpty(object: any): boolean {
+    return !object || object.length === 0;
+  }
+}