+++ test/integration/partner.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 import { ResponseData } from '../../src/infrastructure/utils/Response';
 import { PartnerController } from '../../src/presentation/controllers/PartnerController';
 import DatabaseHelper from './database/databaseHelper';
-import * as mocks from '../mocks/partners';
+import * as mocks from '../mocks';
 const databaseHelper = new DatabaseHelper();
 
 describe('Partner Controller', function () {
@@ -19,15 +19,82 @@ describe('Partner Controller', function () {
     await partnersColletion.deleteMany({});
   });
 
+  context(
+    `knowing that we have two partners that both cover the same area
+    but they are too far from each other (Guaianases and Pinheiros)
+    when given Itaim Paulista location that is closer to Guaianase than Pinheiros`,
+    async function () {
+      it('should return the partner which is located in Guaianases', async function () {
+        const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getDb()));
+
+        const itaimPaulistaLocation = { pathParameters: { longitude: -46.3969364, latitude: -23.4995053 } };
+        await partnerController.post(mocks.partnerGuaianasesMock);
+        await partnerController.post(mocks.partnerPinheirosMock);
+
+        const responsePartner: ResponseData = await partnerController.getPartnerByCoordinates(itaimPaulistaLocation);
+
+        expect(responsePartner.statusCode).to.be.equal(200);
+        expect(JSON.parse(responsePartner.body).tradingName).to.be.equal('Zé Guaianases');
+        expect(JSON.parse(responsePartner.body).ownerName).to.be.equal('Guaianases');
+        expect(JSON.parse(responsePartner.body).document).to.be.equal('60.588.840/0001-49');
+      });
+    }
+  );
+
+  context(
+    `knowing that we have two partners that only one of them covers the client area (Zé Pinheiros partner)
+    also knowing that the nearest partner do not cover the client area (Zé Guaianases parter)`,
+    async function () {
+      it('should return the partner Pinheiros which is located far away from client(Itaim Paulista) but cover the cliet area', async function () {
+        const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getDb()));
+
+        const itaimPaulistaLocation = { pathParameters: { longitude: -46.3969364, latitude: -23.4995053 } };
+        await partnerController.post(mocks.partnerGuaianasesSmallCoveregeAreaMock);
+        await partnerController.post(mocks.partnerPinheirosMock);
+
+        const responsePartner: ResponseData = await partnerController.getPartnerByCoordinates(itaimPaulistaLocation);
+
+        expect(responsePartner.statusCode).to.be.equal(200);
+        expect(JSON.parse(responsePartner.body).tradingName).to.be.equal('Zé Pinheiros');
+        expect(JSON.parse(responsePartner.body).ownerName).to.be.equal('Pinheiros');
+        expect(JSON.parse(responsePartner.body).document).to.be.equal('39.496.795/0001-06');
+      });
+    }
+  );
+
+  context('knowing that we have no partners that covers the client area', async function () {
+    it('should return not found when there is no parter that covers the area', async function () {
+      const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getDb()));
+
+      const itaimPaulistaLocation = { pathParameters: { longitude: -46.3969364, latitude: -23.4995053 } };
+      await partnerController.post(mocks.partnerGuaianasesSmallCoveregeAreaMock);
+
+      const responsePartner: ResponseData = await partnerController.getPartnerByCoordinates(itaimPaulistaLocation);
+
+      expect(responsePartner.statusCode).to.be.equal(404);
+    });
+  });
+
+  context('When try to get the nearest partner ', async function () {
+    it('should return error when mongo is unavailabe', async function () {
+      const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getInvalidDb()));
+      const someLocation = { pathParameters: { longitude: -46.3969364, latitude: -23.4995053 } };
+
+      const result: ResponseData = await partnerController.getPartnerByCoordinates(someLocation);
+
+      expect(result.statusCode).to.be.equal(503);
+    });
+  });
+
   context('when try to save a partner', async function () {
-    it('should save a partner when a new partner is ok', async function () {
+    it('should save a new one when is registered correctly', async function () {
       const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getDb()));
       const result: ResponseData = await partnerController.post(mocks.newPartnerEvent);
 
       expect(result.statusCode).to.be.equal(201);
     });
 
-    it('should return error 409 when try to save a partner with the same document', async function () {
+    it('should not allow to save a partner (error 409) when try to save with the same document', async function () {
       const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getDb()));
       const firstPartner = mocks.newPartnerEvent;
       const secondPartnerWithTheSameDocument = mocks.newPartnerEvent;
@@ -38,7 +105,7 @@ describe('Partner Controller', function () {
       expect(result.statusCode).to.be.equal(409);
     });
 
-    it('should return error 400 when try to save a invalid partner', async function () {
+    it('should not allow to save a partner (error 400) when try to save a invalid partner', async function () {
       const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getInvalidDb()));
       const result: ResponseData = await partnerController.post(mocks.newInvalidAddressPartnerEvent);
 
@@ -71,14 +138,6 @@ describe('Partner Controller', function () {
       const result: ResponseData = await partnerController.get({ pathParameters: { id: 'notExistId' } });
 
       expect(result.statusCode).to.be.equal(404);
-    });
-
-    it('should return not found when the id do not exists in the repository', async function () {
-      const partnerController = new PartnerController(await mocks.partnerService(databaseHelper.getDb()));
-
-      const result: ResponseData = await partnerController.get({});
-
-      expect(result.statusCode).to.be.equal(400);
     });
 
     it('should return error when mongo is unavailabe', async function () {