+++ test/unit/partner.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import { IllegalArgumentException } from '../../src/domain/exceptions/IllegalArgumentException';
 import { Partner } from '../../src/domain/models/partner/partner';
-import * as mocks from '../mocks/partners';
+import * as mocks from '../mocks';
 
 describe('Unit tests for Partner model', () => {
   context('When creating a model partner', () => {
@@ -15,7 +15,6 @@ describe('Unit tests for Partner model', () => {
     it('Should return error when the partner (address) is build incorrectly', () => {
       try {
         Partner.create(null, 'Ze drinks', 'Ze Delivery', mocks.document, mocks.coverageArea, null);
-        
       } catch (error) {
         expect(error).to.be.instanceOf(IllegalArgumentException);
         expect(error.message).to.be.equal('Property address must be defined');