+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.3.0 (2021-01-17)
+
+
+### Features
+
+* find partner method ([38f4762](https://github.com/ricardo-lino/test-ze-delivery/commit/38f4762f486b5dc11a4bb64bf2eff7af77875fba))
+* find partner nearest a coordinate ([3165a16](https://github.com/ricardo-lino/test-ze-delivery/commit/3165a16679292c12220feb1ed36f638712284bbd))
+* save partner method ([892964c](https://github.com/ricardo-lino/test-ze-delivery/commit/892964cdf1ffed3eb67615936030e6de8321c686))
+
 ## 1.2.0 (2021-01-16)